perf(website): fix preload waterfall, add query staleTime, code-split Motion, skip onboarding precache
- Remove build.modulePreload:false so the entry's ~29 vendor chunks are discovered
  via injected modulepreload hints instead of a serial html->entry->vendors
  waterfall (verified: 63 modulepreload links now emitted).
- Add staleTime (60s) to the projects/tags/favorites/goals reference queries to
  stop the focus-refetch storm on the timer page (mutations still invalidate).
- Code-split motion/react out of the eager app-shell chunk via lazy/Suspense with
  the existing plain <span> fallback (verified: motion runtime now only in the
  MotionActiveIndicator chunk, absent from the entry/web-shell chunks).
- Add globIgnores for onboarding PNGs so ~3MB is no longer precached for every
  visitor (verified: 0 onboarding PNGs in the precache manifest).
